bale_bot/requirements.txt
aiohttp>=3.8.1
asyncio>=3.4.3
python-dotenv>=0.20.0 